refactor model error handling module location
