+++ ai_commit_msg/cli/gen_ai_commit_message_handler.py
@@ -1,7 +1,7 @@
 from ai_commit_msg.core.gen_commit_msg import generate_commit_message
 from ai_commit_msg.services.git_service import GitService
 from ai_commit_msg.utils.utils import execute_cli_command
-from ai_commit_msg.services.model_error_handling import AIModelHandlerError
+from ai_commit_msg.utils.model_error_handling import AIModelHandlerError
 from ai_commit_msg.utils.logger import Logger
 
 def gen_ai_commit_message_handler():
