+++ ai_commit_msg/utils/model_error_handling.py
